Improved User-Agent discovery
This commit introduces a number of utilities to interpret the user-agent
strings:
* A new class that introduces a UserAgent object, containing a description of
user browser, version, operating system and a regular expression that can be
used to match against the user agent strings sent with the requests
* The UserActivity class has bee extended with a foreign field to link to the
UserAgent object and a second field that allows us to keep track of the
requested user languages.
* A new script that allows the administrator to load a regular expression based
database of user-agent strings from a fixed website
* A new script that allows the adminstrator to map UserActivities to
UserAgents. This can be done to the whole database or just to a subset which
has not been associated so far - which is at least an order of magnitude
faster.